moving files and applying pr suggestions
